README: leave private key owned by the _snaps user
According to security(8) all private keys must be readable by the owner
only so our hardening to make it readable only for the _snaps user via
it's group permissions and not writable unfortunately makes the daily
security(8) script issue a warning.

Therefore remove the instructions to change ownership of the private key
of the _snaps user to root and just accept the fact that _snaps can
change it's private key.